# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BetRoulette.Api",
    "version": "1.0"
  },
  "paths": {
    "/Bets": {
      "post": {
        "tags": [
          "Bets"
        ],
        "operationId": "CreateBet",
        "parameters": [
          {
            "name": "userId",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Color": {
        "enum": [
          "Black",
          "Red"
        ],
        "type": "string"
      },
      "CreateBetRequest": {
        "required": [
          "amount"
        ],
        "type": "object",
        "properties": {
          "value": {
            "maximum": 36,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "color": {
            "$ref": "#/components/schemas/Color"
          },
          "amount": {
            "maximum": 10000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}